requirements.txt
# Core Dependencies
google-generativeai>=0.8.0
httpx>=0.27.0

# Terminal UI
rich>=13.7.0
prompt-toolkit>=3.0.43

# Configuration
toml>=0.10.2

# File Processing
Pillow>=10.3.0
PyPDF2>=3.0.1

# Utilities
python-dateutil>=2.9.0
aiofiles>=23.2.1
